refactor: (WIP) Folder restructuring
