+++ package.json
@@ -5,8 +5,8 @@
   "main": "dist/server.js",
   "scripts": {
     "build": "tsc",
-    "start": "node dist/server.js",
-    "dev": "ts-node src/server.ts"
+    "start": "node dist/app.js",
+    "dev": "ts-node src/app.ts"
   },
   "author": "",
   "license": "ISC",
